Increaesd the pool_size for SQLAlchemy engine
By default the pool_size is 5. This is the largest number of connections that will be kept persistently in the pool. The pool_size can be set to 0 to indicate no size limit. I determined the sweet spot to be 100 based on the number of projects we have in commit.guru.

Ref: https://dba.stackexchange.com/questions/28060/how-to-determine-the-sweet-spot-between-pool-size-and-database-connections-in-po?rq=1